requirements.txt
asgiref==3.3.1
beautifulsoup4==4.9.3
certifi==2021.5.30
charset-normalizer==2.0.4
cloudinary==1.28.0
Django==3.1.4
django-cloudinary-storage==0.3.0
django-widget-tweaks==1.4.9
djangorestframework==3.13.1
et-xmlfile==1.0.1
idna==3.2
jdcal==1.4.1
MouseInfo==0.1.3
mysql-connector-python==8.0.26
openpyxl==3.0.5
Pillow==8.3.2
psycopg2==2.9.2
PyAutoGUI==0.9.53
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyperclip==1.8.2
PyRect==0.1.4
PyScreeze==0.1.27
PyTweening==1.0.3
pytz==2020.4
pywhatkit==5.1
requests==2.26.0
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.1
urllib3==1.26.6
waitress==2.0.0
whitenoise==5.3.0
wikipedia==1.4.0
